requirements.txt
dnspython
requests
alive-progress